Removed old View Controller
