+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/CardReadersV2/CardReaderSettingsViewModelsOrderedList.swift
@@ -36,7 +36,7 @@ final class CardReaderSettingsViewModelsOrderedList: CardReaderSettingsPrioritiz
                     },
                     knownReaderProvider: knownReaderProvider
                 ),
-                viewIdentifier: "CardReaderSettingsSearchingViewControllerNew"
+                viewIdentifier: "CardReaderSettingsSearchingViewController"
             )
         )
